ecom_project/requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
cssselect2==0.7.0
Django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-model-utils==4.2.0
django-paypal==2.0
django-paypal2==1.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
future==0.18.2
html5lib==1.1
idna==3.4
lxml==4.9.1
oscrypto==1.3.0
paypalrestsdk==1.13.1
pdfkit==1.0.0
Pillow==9.2.0
pycparser==2.21
pydash==5.1.1
pyHanko==0.15.0
pyhanko-certvalidator==0.19.5
PyJWT==2.6.0
PyMySQL==1.0.2
pyOpenSSL==22.1.0
PyPDF2==2.11.1
PyPDF3==1.0.6
python-bidi==0.4.2
python-dotenv==0.21.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.3
stripe==4.2.0
stripe3==0.0.0
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
typing==3.7.4.3
tzdata==2022.5
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
xhtml2pdf==0.2.8
